Refactor: Move requires_requests decorator to base module
Move the _requires_requests decorator from lcsc.py and octopart.py to
base.py to eliminate code duplication. The decorator is now public
(requires_requests) and includes functools.wraps for proper function
metadata preservation.
